@@ -5,39 +5,51 @@
 	"scripts": {
 		"dev": "vite",
 		"build": "tsc && vite build",
-		"serve": "vite preview"
+		"serve": "vite preview",
+		"lint": "eslint src",
+		"lint:fix": "eslint src --fix",
+		"prettier": "prettier src --check",
+		"prettier:fix": "prettier src --write",
+		"clean": "tsc --noEmit && eslint src --fix && prettier src --write"
 	},
 	"dependencies": {
-		"@emotion/react": "^11.7.1",
-		"@emotion/styled": "^11.6.0",
-		"@mui/icons-material": "^5.0.4",
-		"@mui/lab": "^5.0.0-alpha.51",
-		"@mui/material": "^5.0.4",
-		"@reduxjs/toolkit": "^1.7.1",
-		"axios": "^0.22.0",
-		"luxon": "^2.0.2",
-		"no-try": "^3.1.0",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-ga4": "^1.4.1",
+		"@emotion/react": "^11.10.6",
+		"@emotion/styled": "^11.10.6",
+		"@mui/icons-material": "^5.11.9",
+		"@mui/lab": "5.0.0-alpha.120",
+		"@mui/material": "^5.11.9",
+		"@reduxjs/toolkit": "^1.9.2",
+		"axios": "^1.3.3",
+		"luxon": "^3.2.1",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-ga4": "^2.0.0",
 		"react-lazyload": "^3.2.0",
-		"react-redux": "^7.2.5",
-		"react-router-dom": "^6.2.1",
-		"react-use-is-online": "^1.0.4",
-		"spotify-web-api-js": "^1.5.1"
+		"react-redux": "^8.0.5",
+		"react-router-dom": "^6.8.1",
+		"react-use-is-online": "^1.2.1",
+		"spotify-web-api-js": "^1.5.2"
 	},
 	"devDependencies": {
-		"@types/luxon": "^2.0.5",
-		"@types/react": "^17.0.24",
-		"@types/react-dom": "^17.0.9",
-		"@types/react-lazyload": "^3.1.1",
-		"@types/react-redux": "^7.1.18",
-		"@types/react-router-dom": "^5.3.0",
-		"@vitejs/plugin-react": "^1.1.4",
-		"typescript": "^4.4.3",
-		"vite": "^2.7.13",
-		"vite-plugin-pwa": "^0.11.13",
-		"workbox-window": "^6.5.2"
+		"@types/luxon": "^3.2.0",
+		"@types/react": "^18.0.28",
+		"@types/react-dom": "^18.0.11",
+		"@types/react-lazyload": "^3.2.0",
+		"@types/react-redux": "^7.1.25",
+		"@types/react-router-dom": "^5.3.3",
+		"@typescript-eslint/eslint-plugin": "latest",
+		"@typescript-eslint/parser": "latest",
+		"@vitejs/plugin-react": "^3.1.0",
+		"eslint": "latest",
+		"eslint-config-prettier": "latest",
+		"eslint-plugin-react": "latest",
+		"eslint-plugin-simple-import-sort": "latest",
+		"prettier": "latest",
+		"ts-node": "latest",
+		"typescript": "^5.1.6",
+		"vite": "^4.1.1",
+		"vite-plugin-pwa": "^0.14.4",
+		"workbox-window": "^6.5.4"
 	},
 	"browserslist": {
 		"production": [
