@@ -1,18 +1,15 @@
 import React from "react"
-import ReactDOM from "react-dom"
-import { Provider } from "react-redux"
+import { createRoot } from "react-dom/client"
+import { Provider as ReduxProvider } from "react-redux"
 import { BrowserRouter } from "react-router-dom"
 
 import App from "./App"
 import store from "./store"
 
-ReactDOM.render(
-	<React.StrictMode>
-		<Provider store={store}>
-			<BrowserRouter>
-				<App />
-			</BrowserRouter>
-		</Provider>
-	</React.StrictMode>,
-	document.getElementById("root")
+createRoot(document.getElementById("root")!).render(
+	<ReduxProvider store={store}>
+		<BrowserRouter>
+			<App />
+		</BrowserRouter>
+	</ReduxProvider>,
 )
