@@ -1,19 +1,14 @@
-import React from "react"
 import { useNavigate } from "react-router-dom"
 
 import { Card, CardActionArea, CardContent, Typography } from "@mui/material"
 
-interface Props {
-	appearance: iAppearanceCard
-}
-
-const AppearanceCard: React.FC<Props> = (props: Props) => {
-	const { appearance } = props
-
+const AppearanceCard = ({ appearance }: { appearance: iAppearanceCard }) => {
 	const navigate = useNavigate()
 
 	return appearance.condition() ? (
-		<Card sx={{ width: 250, mb: 3 }} onClick={() => navigate(appearance.link)}>
+		<Card
+			sx={{ width: 250, mb: 3 }}
+			onClick={() => navigate(appearance.link)}>
 			<CardActionArea sx={{ height: "100%" }}>
 				<CardContent>
 					<Typography
@@ -22,7 +17,7 @@ const AppearanceCard: React.FC<Props> = (props: Props) => {
 						sx={{
 							width: "fit-content",
 							marginRight: "0.8rem",
-							float: "left"
+							float: "left",
 						}}>
 						{appearance.hash ? "#" : ""}
 						{appearance.condition()}
