@@ -1,39 +1,44 @@
-import React, { useEffect, useState } from "react"
+import React, { useState } from "react"
 
 import { Star, StarBorder } from "@mui/icons-material"
 import {
-	Avatar, Backdrop, Box, Card, CardActionArea, CardMedia, CircularProgress, Dialog, Grid,
-	IconButton, Skeleton, Stack, Tooltip, Typography
+	Avatar,
+	Backdrop,
+	Box,
+	Card,
+	CardActionArea,
+	CardMedia,
+	CircularProgress,
+	Dialog,
+	Grid,
+	IconButton,
+	Skeleton,
+	Stack,
+	Tooltip,
+	Typography,
 } from "@mui/material"
 
-import useAppDispatch from "../../hooks/useAppDispatch"
-import useSpotifyApi from "../../hooks/useSpotifyApi"
-import { set_error } from "../../slices/ErrorSlice"
-import { set_snackbar } from "../../slices/SnackbarSlice"
+import {
+	useAddToMySavedTracksMutation,
+	useGetIsInMySavedTracksQuery,
+	useRemoveFromMySavedTracksMutation,
+} from "../../api/track"
+import useAuthenticated from "../../hooks/useAuthenticated"
 import getDuration from "../../utils/getDuration"
 import AsyncImage from "../AsyncImage"
 import PageIndicator from "../PageIndicator"
 
-interface Props {
-	track?: SpotifyApi.TrackObjectFull
-}
+const TrackDetails = ({ track }: { track?: SpotifyApi.TrackObjectFull }) => {
+	const token = useAuthenticated()
 
-const TrackDetails: React.FC<Props> = (props: Props) => {
-	const { track } = props
+	const { data: isInMySavedTracks } = useGetIsInMySavedTracksQuery(
+		{ ids: [track?.id ?? ""], token },
+		{ skip: !track },
+	)
+	const [addToMySavedTracks] = useAddToMySavedTracksMutation()
+	const [removeFromMySavedTracks] = useRemoveFromMySavedTracksMutation()
 
-	const dispatch = useAppDispatch()
-	const api = useSpotifyApi()
-	const [liked, setLiked] = useState<boolean | null>(null)
 	const [showImage, setShowImage] = useState(false)
-
-	useEffect(() => {
-		if (!api) return
-		if (!track) return
-
-		api.containsMySavedTracks([track.id])
-			.then(res => setLiked(res[0] !== undefined ? res[0] : null))
-			.catch(err => dispatch(set_error(err)))
-	}, [dispatch, api, track])
 
 	const handleTrackOpen = () => {
 		if (track) {
@@ -41,61 +46,19 @@ const TrackDetails: React.FC<Props> = (props: Props) => {
 		}
 	}
 
-	const handleLike = () => {
-		if (api && track) {
-			setLiked(null)
-			api.addToMySavedTracks([track.id])
-				.then(() => {
-					setLiked(true)
-					dispatch(
-						set_snackbar({
-							open: true,
-							message: "Track added to your Liked Tracks",
-							variant: "success"
-						})
-					)
-				})
-				.catch(err => {
-					setLiked(false)
-					dispatch(
-						set_snackbar({
-							open: true,
-							message: "Failed to add Track to your Liked Tracks",
-							variant: "error"
-						})
-					)
-					dispatch(set_error(err))
-				})
-		}
+	const handleLike = async () => {
+		if (!track) return
+
+		await addToMySavedTracks({ ids: [track.id], token })
 	}
 
-	const handleUnlike = () => {
-		if (api && track) {
-			setLiked(null)
-			api.removeFromMySavedTracks([track.id])
-				.then(() => {
-					setLiked(false)
-					dispatch(
-						set_snackbar({
-							open: true,
-							message: "Track removed from your Liked Tracks",
-							variant: "success"
-						})
-					)
-				})
-				.catch(err => {
-					setLiked(true)
-					dispatch(
-						set_snackbar({
-							open: true,
-							message: "Failed to remove Track from your Liked Tracks",
-							variant: "error"
-						})
-					)
-					dispatch(set_error(err))
-				})
-		}
+	const handleUnlike = async () => {
+		if (!track) return
+
+		await removeFromMySavedTracks({ ids: [track.id], token })
 	}
+
+	const liked = isInMySavedTracks?.[0] ?? null
 
 	return (
 		<>
@@ -103,7 +66,9 @@ const TrackDetails: React.FC<Props> = (props: Props) => {
 				sx={{ mt: { xs: 2, sm: 4 }, mb: { sm: 4 } }}
 				container
 				direction={{ xs: "column", sm: "row" }}>
-				<Grid sx={{ mx: { xs: "auto", sm: 2 } }} item>
+				<Grid
+					sx={{ mx: { xs: "auto", sm: 2 } }}
+					item>
 					<AsyncImage
 						src={track?.album.images[0]?.url}
 						skeleton={
@@ -115,7 +80,9 @@ const TrackDetails: React.FC<Props> = (props: Props) => {
 							/>
 						}
 						component={thumbnailUrl => (
-							<Card sx={{ borderRadius: 5 }} onClick={() => setShowImage(true)}>
+							<Card
+								sx={{ borderRadius: 5 }}
+								onClick={() => setShowImage(true)}>
 								<CardActionArea>
 									<CardMedia
 										component="img"
@@ -133,7 +100,7 @@ const TrackDetails: React.FC<Props> = (props: Props) => {
 					sx={{
 						my: { xs: 1, sm: 3 },
 						mx: { xs: "auto", sm: 3 },
-						textAlign: { xs: "center", sm: "start" }
+						textAlign: { xs: "center", sm: "start" },
 					}}
 					item
 					display="flex"
@@ -152,13 +119,28 @@ const TrackDetails: React.FC<Props> = (props: Props) => {
 						</>
 					) : (
 						<>
-							<Skeleton variant="text" width={200} height={45} />
-							<Skeleton variant="text" width={160} height={40} />
-							<Skeleton variant="text" width={80} height={30} />
+							<Skeleton
+								variant="text"
+								width={200}
+								height={45}
+							/>
+							<Skeleton
+								variant="text"
+								width={160}
+								height={40}
+							/>
+							<Skeleton
+								variant="text"
+								width={80}
+								height={30}
+							/>
 						</>
 					)}
 
-					<Stack direction="row" spacing={1} sx={{ mx: { xs: "auto", sm: 0 } }}>
+					<Stack
+						direction="row"
+						spacing={1}
+						sx={{ mx: { xs: "auto", sm: 0 } }}>
 						<Tooltip
 							title={
 								liked === null
@@ -180,7 +162,9 @@ const TrackDetails: React.FC<Props> = (props: Props) => {
 							</IconButton>
 						</Tooltip>
 						<Tooltip title="Open in Spotify">
-							<IconButton sx={{ width: 46 }} onClick={handleTrackOpen}>
+							<IconButton
+								sx={{ width: 46 }}
+								onClick={handleTrackOpen}>
 								<Avatar
 									sx={{ width: 30, height: 30 }}
 									src="/assets/spotify-logo.png"
