@@ -4,15 +4,15 @@ import { useNavigate } from "react-router-dom"
 import { ExpandLess, ExpandMore } from "@mui/icons-material"
 import { Collapse, List, ListItemButton, ListItemIcon, ListItemText } from "@mui/material"
 
-interface Props {
+const DrawerDropdown = ({
+	setOpen,
+	dropdown,
+}: {
 	setOpen: Dispatch<SetStateAction<boolean>>
 	dropdown: iDrawerDropdown
-}
-
-const DrawerDropdown: React.FC<Props> = (props: Props) => {
-	const { setOpen, dropdown } = props
-
+}) => {
 	const navigate = useNavigate()
+
 	const [dropdownOpen, setDropdownOpen] = useState(false)
 
 	const redirect = (url: string) => {
@@ -27,8 +27,13 @@ const DrawerDropdown: React.FC<Props> = (props: Props) => {
 				<ListItemText primary={dropdown.title} />
 				{dropdownOpen ? <ExpandLess /> : <ExpandMore />}
 			</ListItemButton>
-			<Collapse in={dropdownOpen} timeout="auto" unmountOnExit>
-				<List component="div" disablePadding>
+			<Collapse
+				in={dropdownOpen}
+				timeout="auto"
+				unmountOnExit>
+				<List
+					component="div"
+					disablePadding>
 					{dropdown.items.map(item =>
 						item.condition() ? (
 							<ListItemButton
@@ -40,7 +45,7 @@ const DrawerDropdown: React.FC<Props> = (props: Props) => {
 							</ListItemButton>
 						) : (
 							<></>
-						)
+						),
 					)}
 				</List>
 			</Collapse>
