@@ -1,8 +1,17 @@
 import React, { useState } from "react"
 
 import {
-	AccessTime, AccessTimeFilled, Audiotrack, DarkMode, History, Home, LightMode, Login, Logout,
-	Person, Timeline
+	AccessTime,
+	AccessTimeFilled,
+	Audiotrack,
+	DarkMode,
+	History,
+	Home,
+	LightMode,
+	Login,
+	Logout,
+	Person,
+	Timeline,
 } from "@mui/icons-material"
 import MenuIcon from "@mui/icons-material/Menu"
 import { AppBar, Box, IconButton, List, SwipeableDrawer, Toolbar, Typography } from "@mui/material"
@@ -11,8 +20,8 @@ import useAppSelector from "../../hooks/useAppSelector"
 import DrawerDropdown from "./DrawerDropdown"
 import DrawerItem from "./DrawerItem"
 
-const Navigator: React.FC = () => {
-	const access_token = useAppSelector(state => state.access_token)
+const Navigator = () => {
+	const token = useAppSelector(state => state.token)
 	const theme = useAppSelector(state => state.theme)
 	const [open, setOpen] = useState(false)
 
@@ -22,14 +31,14 @@ const Navigator: React.FC = () => {
 			icon: <Home />,
 			title: "Home",
 			url: "/",
-			condition: () => true
+			condition: () => true,
 		},
 		{
 			id: "nav-login",
 			icon: <Login />,
 			title: "Login",
 			url: "/login",
-			condition: () => !access_token
+			condition: () => !token,
 		},
 		{
 			id: "nav-top-tracks",
@@ -42,24 +51,24 @@ const Navigator: React.FC = () => {
 					icon: <AccessTime />,
 					title: "Last Month",
 					url: "/top-tracks/short-term",
-					condition: () => !!access_token
+					condition: () => !!token,
 				},
 				{
 					id: "nav-top-tracks-medium-term",
 					icon: <AccessTimeFilled />,
 					title: "Last 6 Months",
 					url: "/top-tracks/medium-term",
-					condition: () => !!access_token
+					condition: () => !!token,
 				},
 				{
 					id: "nav-top-tracks-long-term",
 					icon: <Timeline />,
 					title: "All Time",
 					url: "/top-tracks/long-term",
-					condition: () => !!access_token
-				}
+					condition: () => !!token,
+				},
 			],
-			condition: () => !!access_token
+			condition: () => !!token,
 		},
 		{
 			id: "nav-top-artists",
@@ -72,58 +81,60 @@ const Navigator: React.FC = () => {
 					icon: <AccessTime />,
 					title: "Last Month",
 					url: "/top-artists/short-term",
-					condition: () => !!access_token
+					condition: () => !!token,
 				},
 				{
 					id: "nav-top-artists-medium-term",
 					icon: <AccessTimeFilled />,
 					title: "Last 6 Months",
 					url: "/top-artists/medium-term",
-					condition: () => !!access_token
+					condition: () => !!token,
 				},
 				{
 					id: "nav-top-artists-long-term",
 					icon: <Timeline />,
 					title: "All Time",
 					url: "/top-artists/long-term",
-					condition: () => !!access_token
-				}
+					condition: () => !!token,
+				},
 			],
-			condition: () => !!access_token
+			condition: () => !!token,
 		},
 		{
 			id: "nav-recents",
 			icon: <History />,
 			title: "Recently Played",
 			url: "/recents",
-			condition: () => !!access_token
+			condition: () => !!token,
 		},
 		{
 			id: "nav-to-light",
 			icon: <LightMode />,
 			title: "Light Mode",
 			url: "/light",
-			condition: () => theme === "dark"
+			condition: () => theme === "dark",
 		},
 		{
 			id: "nav-to-dark",
 			icon: <DarkMode />,
 			title: "Dark Mode",
 			url: "/dark",
-			condition: () => theme === "light"
+			condition: () => theme === "light",
 		},
 		{
 			id: "nav-logout",
 			icon: <Logout />,
 			title: "Logout",
 			url: "/logout",
-			condition: () => !!access_token
-		}
+			condition: () => !!token,
+		},
 	]
 
 	return (
 		<>
-			<AppBar sx={{ bgcolor: "primary.main" }} position="relative">
+			<AppBar
+				sx={{ bgcolor: "primary.main" }}
+				position="relative">
 				<Toolbar>
 					<IconButton
 						onClick={() => setOpen(true)}
@@ -142,19 +153,31 @@ const Navigator: React.FC = () => {
 				open={open}
 				onOpen={() => setOpen(true)}
 				onClose={() => setOpen(false)}>
-				<Box sx={{ width: 250 }} role="presentation">
+				<Box
+					sx={{ width: 250 }}
+					role="presentation">
 					<List
 						subheader={
-							<Typography sx={{ px: 3, py: 2 }} variant="h4">
+							<Typography
+								sx={{ px: 3, py: 2 }}
+								variant="h4">
 								Statify
 							</Typography>
 						}>
 						{data.map(el =>
 							"drop" in el ? (
-								<DrawerDropdown key={el.id} setOpen={setOpen} dropdown={el} />
+								<DrawerDropdown
+									key={el.id}
+									setOpen={setOpen}
+									dropdown={el}
+								/>
 							) : (
-								<DrawerItem key={el.id} setOpen={setOpen} item={el} />
-							)
+								<DrawerItem
+									key={el.id}
+									setOpen={setOpen}
+									item={el}
+								/>
+							),
 						)}
 					</List>
 				</Box>
