@@ -2,15 +2,16 @@ import { Alert, Fade, Snackbar as MUISnackbar } from "@mui/material"
 
 import useAppDispatch from "../../hooks/useAppDispatch"
 import useAppSelector from "../../hooks/useAppSelector"
-import { clear_snackbar } from "../../slices/SnackbarSlice"
+import { clearSnackbar } from "../../slices/SnackbarSlice"
 
 const InfoSnackbar = () => {
-	const dispatch = useAppDispatch()
 	const snackbar = useAppSelector(state => state.snackbar)
+	const dispatch = useAppDispatch()
 
 	const handleClose = () => {
-		dispatch(clear_snackbar())
+		dispatch(clearSnackbar())
 	}
+
 	return (
 		<MUISnackbar
 			open={snackbar.open}
