@@ -1,22 +1,28 @@
-import React from "react"
 import LazyLoad from "react-lazyload"
 import { useNavigate } from "react-router-dom"
 
 import {
-	Avatar, Card, CardActionArea, ListItem, ListItemAvatar, ListItemText, Skeleton, Stack
+	Avatar,
+	Card,
+	CardActionArea,
+	ListItem,
+	ListItemAvatar,
+	ListItemText,
+	Skeleton,
+	Stack,
 } from "@mui/material"
 
 import AsyncImage from "./AsyncImage"
 
-interface Props {
+const Track = ({
+	track,
+	album,
+	i,
+}: {
 	track?: SpotifyApi.TrackObjectFull | SpotifyApi.TrackObjectSimplified
 	album?: SpotifyApi.AlbumObjectSimplified
 	i: number
-}
-
-const Track: React.FC<Props> = (props: Props) => {
-	const { track, album, i } = props
-
+}) => {
 	const navigate = useNavigate()
 
 	const handleTrackClick = (track?: SpotifyApi.TrackObjectSimplified) => {
@@ -27,7 +33,10 @@ const Track: React.FC<Props> = (props: Props) => {
 
 	return (
 		<LazyLoad height={72}>
-			<Card sx={{ my: 1 }} key={i} onClick={() => handleTrackClick(track)}>
+			<Card
+				sx={{ my: 1 }}
+				key={i}
+				onClick={() => handleTrackClick(track)}>
 				<CardActionArea>
 					<ListItem>
 						<ListItemAvatar>
@@ -36,9 +45,18 @@ const Track: React.FC<Props> = (props: Props) => {
 									(track && "album" in track ? track.album : album)?.images[0]
 										?.url
 								}
-								skeleton={<Skeleton variant="circular" width={45} height={45} />}
+								skeleton={
+									<Skeleton
+										variant="circular"
+										width={45}
+										height={45}
+									/>
+								}
 								component={thumbnailUrl => (
-									<Avatar sx={{ width: 45, height: 45 }} src={thumbnailUrl} />
+									<Avatar
+										sx={{ width: 45, height: 45 }}
+										src={thumbnailUrl}
+									/>
 								)}
 							/>
 						</ListItemAvatar>
@@ -49,8 +67,16 @@ const Track: React.FC<Props> = (props: Props) => {
 							/>
 						) : (
 							<Stack my="6px">
-								<Skeleton variant="text" width={200} height={24} />
-								<Skeleton variant="text" width={160} height={20} />
+								<Skeleton
+									variant="text"
+									width={200}
+									height={24}
+								/>
+								<Skeleton
+									variant="text"
+									width={160}
+									height={20}
+								/>
 							</Stack>
 						)}
 					</ListItem>
