@@ -1,6 +1,6 @@
 import useAppSelector from "./useAppSelector"
 
-const useThemeValue = <T extends any>(dark: T, light: T): T => {
+const useThemeValue = <T>(dark: T, light: T): T => {
 	const theme = useAppSelector(state => state.theme)
 
 	return theme === "dark" ? dark : light
