@@ -2,14 +2,15 @@ import React, { useEffect } from "react"
 import { useNavigate } from "react-router-dom"
 
 import useAppDispatch from "../hooks/useAppDispatch"
-import { set_theme } from "../slices/ThemeSlice"
+import { setTheme } from "../slices/ThemeSlice"
 
-const Light: React.FC = () => {
-	const dispatch = useAppDispatch()
+const Light = () => {
 	const navigate = useNavigate()
 
+	const dispatch = useAppDispatch()
+
 	useEffect(() => {
-		dispatch(set_theme("light"))
+		dispatch(setTheme("light"))
 		navigate(-1)
 	}, [dispatch, navigate])
 
