@@ -3,38 +3,57 @@ import { useLocation, useNavigate } from "react-router-dom"
 
 import { TabContext } from "@mui/lab"
 import {
-	Box, Card, CardContent, Container, LinearProgress, List, Paper, Tab, Table, TableBody,
-	TableCell, TableContainer, TableHead, TableRow, Tabs, Typography, useMediaQuery, useTheme
+	Box,
+	Card,
+	CardContent,
+	Container,
+	LinearProgress,
+	List,
+	Paper,
+	Tab,
+	Table,
+	TableBody,
+	TableCell,
+	TableContainer,
+	TableHead,
+	TableRow,
+	Tabs,
+	Typography,
+	useMediaQuery,
+	useTheme,
 } from "@mui/material"
 
 import { tabs } from "../App"
 import TopTrackItem from "../components/Items/TopTrackItem"
-import useAppSelector from "../hooks/useAppSelector"
-import useAuthenticated from "../hooks/useAthenticated"
+import useAuthenticated from "../hooks/useAuthenticated"
+import useGetFullTopTracksQuery from "../hooks/useGetFullTopTracksQuery"
 
-const TopTracks: React.FC = () => {
-	const tracks = useAppSelector(state => state.statistics.tracks)
-	const location = useLocation()
+const TopTracks = () => {
+	const token = useAuthenticated()
+
 	const navigate = useNavigate()
-	const [term, setTerm] = useState<"" | "short_term" | "medium_term" | "long_term">("")
+	const location = useLocation()
 	const theme = useTheme()
-	const smallScreen = useMediaQuery(theme.breakpoints.down("lg")) // in wrong order but needs theme
+	const smallScreen = useMediaQuery(theme.breakpoints.down("lg"))
 
-	useAuthenticated()
+	const [term, setTerm] = useState<"" | "short_term" | "medium_term" | "long_term">("")
+	const { data: topTracks } = useGetFullTopTracksQuery(
+		{
+			time_range: term as "short_term" | "medium_term" | "long_term",
+			token,
+		},
+		{ skip: !term },
+	)
 
 	useEffect(() => {
 		const parts = location.pathname.split("/")
 		setTerm(
-			parts[parts.length - 1]!.replace("-", "_") as "short_term" | "medium_term" | "long_term"
+			parts[parts.length - 1]!.replace("-", "_") as
+				| "short_term"
+				| "medium_term"
+				| "long_term",
 		)
 	}, [location.pathname])
-
-	useEffect(() => {
-		if (Object.values(tracks).includes(null)) {
-			sessionStorage.setItem("redirect", location.pathname)
-			navigate("/login")
-		}
-	}, [navigate, location.pathname, tracks])
 
 	return term ? (
 		<TabContext value={term}>
@@ -43,15 +62,26 @@ const TopTracks: React.FC = () => {
 					value={term}
 					onChange={(e, tab) => navigate("../" + tab.replace("_", "-"))}
 					centered>
-					<Tab label="Last Month" value="short_term" />
-					<Tab label="Last 6 Months" value="medium_term" />
-					<Tab label="All Time" value="long_term" />
+					<Tab
+						label="Last Month"
+						value="short_term"
+					/>
+					<Tab
+						label="Last 6 Months"
+						value="medium_term"
+					/>
+					<Tab
+						label="All Time"
+						value="long_term"
+					/>
 				</Tabs>
 				<Container sx={{ mt: 3 }}>
 					<Card>
 						<CardContent>
 							<Typography variant="h4">Top Tracks</Typography>
-							<Typography variant="h6" gutterBottom>
+							<Typography
+								variant="h6"
+								gutterBottom>
 								{tabs.find(t => t.term === term)?.description}
 							</Typography>
 							<Typography variant="body1">
@@ -62,11 +92,11 @@ const TopTracks: React.FC = () => {
 					<Card sx={{ my: 3 }}>
 						{smallScreen ? (
 							<List>
-								{(tracks[term] || Array(5).fill(undefined)).map((track, i) => (
+								{(topTracks ?? Array(5).fill(undefined)).map((t, i) => (
 									<TopTrackItem
 										key={i}
 										smallScreen={smallScreen}
-										track={track}
+										track={t}
 										i={i}
 									/>
 								))}
@@ -84,16 +114,14 @@ const TopTracks: React.FC = () => {
 										</TableRow>
 									</TableHead>
 									<TableBody>
-										{(tracks[term] || Array(5).fill(undefined)).map(
-											(track, i) => (
-												<TopTrackItem
-													key={i}
-													smallScreen={smallScreen}
-													track={track}
-													i={i}
-												/>
-											)
-										)}
+										{(topTracks ?? Array(5).fill(undefined)).map((t, i) => (
+											<TopTrackItem
+												key={i}
+												smallScreen={smallScreen}
+												track={t}
+												i={i}
+											/>
+										))}
 									</TableBody>
 								</Table>
 							</TableContainer>
