@@ -1,5 +1,3 @@
-import { useTry as _useTry } from "no-try"
-
 import { createSlice, PayloadAction } from "@reduxjs/toolkit"
 
 type iErrorData = {
@@ -11,34 +9,34 @@ const slice = createSlice({
 	name: "error",
 	initialState: null as iErrorData,
 	reducers: {
-		set_error: (state, action: PayloadAction<Error | null>) => {
+		setError: (state, action: PayloadAction<Error | null>) => {
 			if (action.payload) {
 				const response_str = (action.payload as any).response
 				if (response_str) {
-					const [err, response] = _useTry(() => JSON.parse(response_str))
-					if (err) {
-						return {
-							name: "Error",
-							message: response
-						}
-					} else {
+					try {
+						const response = JSON.parse(response_str)
 						return {
 							name: `HTTP Error ${response.error.status}`,
-							message: response.error.message
+							message: response.error.message,
+						}
+					} catch (err) {
+						return {
+							name: "Error",
+							message: (err as Error).message,
 						}
 					}
 				} else {
 					return {
 						name: action.payload.name,
-						message: action.payload.message
+						message: action.payload.message,
 					}
 				}
 			} else {
 				return null
 			}
-		}
-	}
+		},
+	},
 })
 
-export const { set_error } = slice.actions
+export const { setError } = slice.actions
 export default slice.reducer
