@@ -2,14 +2,14 @@ import { createTheme } from "@mui/material"
 
 const palette = {
 	success: {
-		main: "#1db954"
+		main: "#1db954",
 	},
 	error: {
-		main: "#e22134"
+		main: "#e22134",
 	},
 	warning: {
-		main: "#ff5722"
-	}
+		main: "#ff5722",
+	},
 }
 
 const light = createTheme({
@@ -17,12 +17,12 @@ const light = createTheme({
 		...palette,
 		primary: {
 			main: "#1ED760",
-			contrastText: "#111111"
+			contrastText: "#111111",
 		},
 		background: {
-			default: "#EEEEEE"
-		}
-	}
+			default: "#EEEEEE",
+		},
+	},
 })
 
 const dark = createTheme({
@@ -30,13 +30,13 @@ const dark = createTheme({
 		...palette,
 		primary: {
 			main: "#25b359",
-			contrastText: "#DDDDDD"
+			contrastText: "#DDDDDD",
 		},
 		mode: "dark",
 		background: {
-			default: "#20262D"
-		}
-	}
+			default: "#20262D",
+		},
+	},
 })
 
 export { light, dark }
